+++ wgpu-core/src/command/compute_command.rs
@@ -17,20 +17,20 @@ pub enum ComputeCommand<R: ReferenceType> {
 
     SetPipeline(R::ComputePipeline),
 
-    /// Set a range of push constants to values stored in `push_constant_data`.
-    SetPushConstant {
-        /// The byte offset within the push constant storage to write to. This
+    /// Set a range of immediates to values stored in `immediates_data`.
+    SetImmediate {
+        /// The byte offset within the immediate data storage to write to. This
         /// must be a multiple of four.
         offset: u32,
 
         /// The number of bytes to write. This must be a multiple of four.
         size_bytes: u32,
 
-        /// Index in `push_constant_data` of the start of the data
+        /// Index in `immediates_data` of the start of the data
         /// to be written.
         ///
         /// Note: this is not a byte offset like `offset`. Rather, it is the
-        /// index of the first `u32` element in `push_constant_data` to read.
+        /// index of the first `u32` element in `immediates_data` to read.
         values_offset: u32,
     },