+++ wgpu-hal/examples/halmark/main.rs
@@ -242,7 +242,7 @@ impl<A: hal::Api> Example<A> {
             label: None,
             flags: hal::PipelineLayoutFlags::empty(),
             bind_group_layouts: &[&global_group_layout, &local_group_layout],
-            push_constant_ranges: &[],
+            immediates_ranges: &[],
         };
         let pipeline_layout = unsafe {
             device